+++ src/main/java/com/korea/it/shopping/product/repo/ProductRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
     List<Product> findByProductNameContainingIgnoreCase(String query); // 대소문자 구분 없이 이름에 검색어가 포함된 상품 검색
-    Optional<Product> findByProductId(Long id);
+    Optional<Product> findByProductId(Integer id);
 }
